Uses explicit UTF-8 decoding for 'sqlFromFile'
The built-in 'readFile' function uses the system locale when determining
how to convert the input bytes to a Haskell 'String', which can be a
source of annoying and difficult to track down issues with respect to
character encoding.

Instead, it's safer to make the assumption that UTF-8 is the most
commonly used encoding scheme and reject all non-UTF-8 input as invalid
at compile-time whenever the 'sqlFromFile' splice is invoked.