Add cancelable option
When this option is enabled, asynchronous exceptions received
while waiting on a connection will be intercepted, and cause
a cancel message to be sent to the server.

If the server is responsive, this will cause blocking FFI
calls to be interrupted promptly, instead of always waiting
for the command to be handled.

The intent is to allow client code to interrupt long-running
queries using System.Timeout.timeout.