fix: don't throw cancel handle errors on invalid connections
When we can't get a cancel handle, we used to throw an error. This is
confusing and misleading for the user, as there is no error/issue on the
cancellation feature, rather that the connection is invalid.

This fixes it to not throw the error "failed to allocate cancel handle".

We can't get a cancel handle when the connection is invalid. So we
don't setup the cancellation, because the query is going to fail anyway
on this invalid connection. If we throw an error here, it becomes
confusing for the user. An appropriate way to handle it would be to throw
an invalid connection error here. But we don't have any details on why
the connection is invalid. If we let postgres throw the error, it will
have the actual error message/reason for the invalid connection. Hence,
it makes sense to not setup the cancellation here and try to run the
query, which will fail with proper error message from Postgres.